auto-configuration of git ignore revs file path
